+++ MilkCRM-backEnd/src/main/java/com/ekaryagin/milkcrm/controller/ProductGroupController.java
@@ -3,12 +3,17 @@ package com.ekaryagin.milkcrm.controller;
 import com.ekaryagin.milkcrm.dto.Mapper;
 import com.ekaryagin.milkcrm.dto.ProductGroupDTO;
 import com.ekaryagin.milkcrm.dto.ProductGroupDTOext;
+import com.ekaryagin.milkcrm.entity.employee.Role;
+import com.ekaryagin.milkcrm.entity.employee.User;
 import com.ekaryagin.milkcrm.entity.products.ProductGroup;
+import com.ekaryagin.milkcrm.security.jwt.JwtTokenProvider;
 import com.ekaryagin.milkcrm.service.ProductGroupService;
+import com.ekaryagin.milkcrm.service.UserService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpServletRequest;
 import java.util.List;
 
 @RestController
@@ -16,14 +21,35 @@ public class ProductGroupController {
 
     private final ProductGroupService productGroupService;
     private final Mapper mapper;
+    private final UserService userService;
+    private final HttpServletRequest request;
+    private final JwtTokenProvider jwtTokenProvider;
 
-    public ProductGroupController(ProductGroupService productGroupService, Mapper mapper) {
+    public ProductGroupController(ProductGroupService productGroupService,
+                                  Mapper mapper,
+                                  UserService userService,
+                                  HttpServletRequest request,
+                                  JwtTokenProvider jwtTokenProvider) {
         this.productGroupService = productGroupService;
         this.mapper = mapper;
+        this.userService = userService;
+        this.request = request;
+        this.jwtTokenProvider = jwtTokenProvider;
+    }
+
+    private User getAuthorizedUser(){
+        String token = request.getHeader("authorization").substring(7);
+        return userService.getUser(jwtTokenProvider.getUsername(token));
     }
 
     @GetMapping(value = "/product-groups")
     public ResponseEntity<List<ProductGroupDTO>> readProductGroups() {
+
+        User user = getAuthorizedUser();
+        if (user.getRole() != Role.ADMIN && user.getRole() != Role.MANAGER){
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+
         List<ProductGroup> productGroups = productGroupService.getProductGroups();
 
         return productGroups != null && !productGroups.isEmpty()
@@ -59,6 +85,11 @@ public class ProductGroupController {
     @PostMapping(value = "/product-groups")
     public ResponseEntity<String> newProductGroup (@RequestBody ProductGroupDTOext body){
 
+        User user = getAuthorizedUser();
+        if (user.getRole() != Role.ADMIN && user.getRole() != Role.MANAGER){
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+
         switch (productGroupService.addNewGroup(mapper.mapFromProductGroupDTOext(body))){
             case (0):
                 return new ResponseEntity<>(HttpStatus.OK);
@@ -80,6 +111,11 @@ public class ProductGroupController {
 
     @PutMapping(value = "/product-groups")
     public ResponseEntity<String> saveProductGroup (@RequestBody ProductGroupDTOext body){
+
+        User user = getAuthorizedUser();
+        if (user.getRole() != Role.ADMIN && user.getRole() != Role.MANAGER){
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
 
         switch (productGroupService.saveGroup(mapper.mapFromProductGroupDTOext(body))){
             case (0):
@@ -104,6 +140,12 @@ public class ProductGroupController {
 
     @DeleteMapping(value = "/product-groups/{id}")
     public ResponseEntity<String> deleteProductGroup (@PathVariable(name = "id") long id) {
+
+        User user = getAuthorizedUser();
+        if (user.getRole() != Role.ADMIN && user.getRole() != Role.MANAGER){
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+
         if (productGroupService.deleteProductGroup(id)) {
             return new ResponseEntity<>(HttpStatus.OK);
         }
