+++ MilkCRM-backEnd/src/main/java/com/ekaryagin/milkcrm/controller/SellerController.java
@@ -2,13 +2,17 @@ package com.ekaryagin.milkcrm.controller;
 
 import com.ekaryagin.milkcrm.dto.Mapper;
 import com.ekaryagin.milkcrm.dto.SellerDTO;
+import com.ekaryagin.milkcrm.entity.employee.Role;
 import com.ekaryagin.milkcrm.entity.employee.Seller;
+import com.ekaryagin.milkcrm.entity.employee.User;
+import com.ekaryagin.milkcrm.security.jwt.JwtTokenProvider;
 import com.ekaryagin.milkcrm.service.RegionService;
 import com.ekaryagin.milkcrm.service.UserService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpServletRequest;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -17,15 +21,26 @@ public class SellerController {
     private final UserService userService;
     private final RegionService regionService;
     private final Mapper mapper;
+    private final HttpServletRequest request;
+    private final JwtTokenProvider jwtTokenProvider;
 
-    public SellerController(UserService userService, RegionService regionService, Mapper mapper) {
+    public SellerController(UserService userService,
+                            RegionService regionService,
+                            Mapper mapper,
+                            HttpServletRequest request,
+                            JwtTokenProvider jwtTokenProvider) {
         this.userService = userService;
         this.regionService = regionService;
         this.mapper = mapper;
+        this.request = request;
+        this.jwtTokenProvider = jwtTokenProvider;
     }
 
+    private User getAuthorizedUser(){
+        String token = request.getHeader("authorization").substring(7);
+        return userService.getUser(jwtTokenProvider.getUsername(token));
+    }
 
-    // Get all sellers both wholesale and retail
     @GetMapping(value = "/sellers")
     public ResponseEntity<List<SellerDTO>> readSellers() {
         List<Seller> sellers = userService.readAllSellers();
@@ -35,7 +50,6 @@ public class SellerController {
                 : new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
-    // Get a seller by id
     @GetMapping(value = "/sellers/{id}")
     public ResponseEntity<SellerDTO> readSeller(@PathVariable(name = "id") long id) {
         Seller seller = userService.readSeller(id);
@@ -45,7 +59,6 @@ public class SellerController {
                 : new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
-    // Get all retail sellers
     @GetMapping(value = "/sellers/retail-sellers")
     public ResponseEntity<List<SellerDTO>> readRetailSellers() {
         List<Seller> sellers = userService.readAllRetailSellers();
@@ -55,7 +68,6 @@ public class SellerController {
                 : new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
-    // Get all dealers
     @GetMapping(value = "/sellers/dealers")
     public ResponseEntity<List<SellerDTO>> readDealers() {
         List<Seller> sellers = userService.readAllDealers();
@@ -65,8 +77,6 @@ public class SellerController {
                 : new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
-
-    // Get all sellers from region(by id)
     @GetMapping(value = "/region/{id}/sellers")
     public ResponseEntity<List<SellerDTO>> readSellersFromRegion(@PathVariable(name = "id") long id) {
         List<Seller> sellers = userService.readAllSellersFromRegion(regionService.readRegionById(id));
@@ -76,7 +86,6 @@ public class SellerController {
                 : new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
-    //Get all retail sellers from region(by id)
     @GetMapping(value = "/region/{id}/sellers/retail-sellers")
     public ResponseEntity<List<SellerDTO>> readRetailSellersFromRegion(@PathVariable(name = "id") long id) {
         List<Seller> sellers = userService.readAllRetailSellersFromRegion(regionService.readRegionById(id));
@@ -86,7 +95,6 @@ public class SellerController {
                 : new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
-    //Get all dealers from region(by id)
     @GetMapping(value = "/region/{id}/sellers/dealers")
     public ResponseEntity<List<SellerDTO>> readDealersFromRegion(@PathVariable(name = "id") long id) {
         List<Seller> sellers = userService.readAllDealersFromRegion(regionService.readRegionById(id));
@@ -96,7 +104,6 @@ public class SellerController {
                 : new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
-    //Get all sellers from shop(by id)
     @GetMapping(value = "/shop/{id}/sellers")
     public ResponseEntity<List<SellerDTO>> readSellersFromShop(@PathVariable(name = "id") long id) {
         List<Seller> sellers = new ArrayList<>(userService.readAllSellersFromShop(id));
@@ -106,9 +113,14 @@ public class SellerController {
                 : new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
-    // Post new seller, 422 if client with this id already exists
     @PostMapping(value = "/sellers")
     public ResponseEntity<String> newSeller(@RequestBody SellerDTO body){
+
+        User user = getAuthorizedUser();
+        if (user.getRole() != Role.ADMIN && user.getRole() != Role.MANAGER
+                && !body.getUsername().equals(user.getUsername())){
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
 
         switch (userService.addUser(mapper.mapFromSellerDTO(body))){
             case (0):
@@ -131,6 +143,11 @@ public class SellerController {
 
     @PutMapping(value = "/sellers")
     public ResponseEntity<String> saveSeller(@RequestBody SellerDTO body){
+
+        User user = getAuthorizedUser();
+        if (user.getRole() != Role.ADMIN && user.getRole() != Role.MANAGER){
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
 
         switch (userService.saveUser(mapper.mapFromSellerDTO(body))){
             case (0):
@@ -157,9 +174,16 @@ public class SellerController {
 
     @DeleteMapping(value = "/sellers/{id}")
     public ResponseEntity<String> deleteSeller(@PathVariable(name = "id") long id){
+
+        User user = getAuthorizedUser();
+        if (user.getRole() != Role.ADMIN && user.getRole() != Role.MANAGER){
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+
         if (userService.deleteSeller(id)) {
             return new ResponseEntity<>(HttpStatus.OK);
         }
+
         return new ResponseEntity<>("This seller cannot be deleted", HttpStatus.EXPECTATION_FAILED);
     }
